refactor: 회원가입, 로그인 상태 확인 API memberDomain으로 이동
