--- module-api/src/main/java/kernel/jdon/moduleapi/domain/auth/core/AuthService.java
@@ -1,5 +1,0 @@
-package kernel.jdon.moduleapi.domain.auth.core;
-
-public interface AuthService {
-	AuthInfo.RegisterResponse register(AuthCommand.RegisterRequest command);
-}
