+++ module-api/src/main/java/kernel/jdon/moduleapi/domain/member/presentation/MemberController.java
@@ -1,5 +1,7 @@
 package kernel.jdon.moduleapi.domain.member.presentation;
 
+import java.net.URI;
+
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -51,5 +53,31 @@ public class MemberController {
 		memberFacade.checkNicknameDuplicate(command);
 
 		return ResponseEntity.noContent().build();
+	}
+
+	@PostMapping("/api/v1/register")
+	public ResponseEntity<CommonResponse<MemberDto.RegisterResponse>> register(
+		@RequestBody @Valid final MemberDto.RegisterRequest request) {
+		final MemberCommand.RegisterRequest command = memberDtoMapper.of(request);
+		final MemberInfo.RegisterResponse info = memberFacade.register(command);
+		final MemberDto.RegisterResponse response = memberDtoMapper.of(info);
+		final URI uri = URI.create("/api/v1/member/" + response.getMemberId());
+
+		return ResponseEntity.created(uri).body(CommonResponse.of(response));
+
+	}
+
+	@GetMapping("/api/v1/authenticate")
+	public ResponseEntity<CommonResponse<MemberDto.AuthenticateResponse>> authenticate(
+		@LoginUser final SessionUserInfo sessionUser) {
+		boolean isLoginUser = false;
+		Long memberId = null;
+		if (null != sessionUser) {
+			isLoginUser = true;
+			memberId = sessionUser.getId();
+		}
+		final MemberDto.AuthenticateResponse response = MemberDto.AuthenticateResponse.of(isLoginUser, memberId);
+
+		return ResponseEntity.ok().body(CommonResponse.of(response));
 	}
 }
